Added low percent error version

+++ Mechatronics/NONEABISoftware/MicrocontrollerArduinoCode/ArduinoVersion/hola.h
@@ -97,7 +97,7 @@ char get_char(void);
 void initEncoder(void);
 void ext_interruptConfig(void);
 void timer1Config(void);
-int8_t getPositionValues(int *target);
+int8_t getPositionValues(uint16_t *target);
 
 #define RN4020_MAX_UUID_LEN_BYTES (128 / 8)
 
@@ -256,13 +256,13 @@ void setPrivateCharacteristic(char *service,
             uint8_t dataSize,
             uint8_t securityFlag);
 void setName(char *string);//No >  6 bytes when using private service
-void writeOneOneByteCharacteristic(int handle, int8_t value0);
-void writeTenTwoBytesCharacteristic(int handle, int16_t value0,
-            int16_t value1, int16_t value2,
-            int16_t value3, int16_t value4,
-            int16_t value5, int16_t value6,
-            int16_t value7, int16_t value8,
-            int16_t value9);
+void writeOneOneByteCharacteristic(unsigned int handle, uint8_t value0);
+void writeTenTwoBytesCharacteristic(unsigned int handle, uint16_t value0,
+            uint16_t value1, uint16_t value2,
+            uint16_t value3, uint16_t value4,
+            uint16_t value5, uint16_t value6,
+            uint16_t value7, uint16_t value8,
+            uint16_t value9);
 void turnOffSubscription(int handle);
 
 //-------------- USART PARSING FUNCTIONS----------
