+++ Mechatronics/NONEABISoftware/MicrocontrollerCode/Src_Modules/MicroC_Application/src/ENCODER/NONCOMUNICATION_PERIPHEREALS/Encoder.h
@@ -3,11 +3,11 @@
 
 #include <inttypes.h>
 
-extern volatile int16_t encoderCounter;
+extern volatile uint16_t encoderCounter;
 
 void initEncoder(void);
 void ext_interruptConfig(void);
 void timer1Config(void);
-int8_t getPositionValues(int *target);
+int8_t getPositionValues(uint16_t *target);
 
 #endif
